PI-2342 Enable trace propagation across SNS/SQS (#173)
* PI-2342 Enable trace propagation across SNS/SQS

This PR adds the OpenTelemetry trace context into the MessageAttributes before publishing to a topic or sending to a queue. It also adds an interceptor to extract the trace context, if present, and configure the current span on receiving a message.

The result is a single joined-up transaction in Application Insights, allowing you to trace message processing back to the message source. It also enables a richer dependency graph, as we can now track which services interact with each other over the SNS/SQS message boundary.

* Remove dependency on opentelemetry-instrumentation-annotations

* Updated tests to check for exported spans

* Remove dependency version for opentelemetry-extension-kotlin

* Bump version to 4.1.0

* Add -beta suffix to version